+++ src/main/java/svenhjol/charmony/api/StoneChestLockMenuData.java
@@ -16,5 +16,5 @@ public class StoneChestLockMenuData {
     public BlockPos pos;
     public ContainerData data;
     public long seed;
-    public double difficultyAmplifier;
+    public int difficultyAmplifier;
 }
